% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tag.R
\name{get_engine_fn}
\alias{get_engine_fn}
\title{Get engine-specific function string}
\usage{
get_engine_fn(engine, suffix, out)
}
\arguments{
\item{engine}{logdec engine to use, i.e. \code{cli}}

\item{suffix}{Suffix of engine function, or an empty character}

\item{out}{The user-inputted output}
}
\value{
An unparsed string for an engine-specific function
}
\description{
Get engine-specific function string
}
\seealso{
\code{parse_tag}
}
\keyword{internal}
