% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tag.R
\name{parse_tag}
\alias{parse_tag}
\title{Parse logdec comment tag}
\usage{
parse_tag(block, engine)
}
\arguments{
\item{block}{A logdec comment tag as a \code{character}}

\item{engine}{The logdec engine to use, passed to \code{get_engine_fn}.}
}
\value{
An unparsed string for an engine-specific function
}
\description{
Parse logdec comment tag
}
\seealso{
\code{get_engine_fn}
}
\keyword{internal}
